backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.7.8
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0
google-auth==2.35.0

pytest==8.3.5
